Update README to 1.0.1
Version 1.0.1 has an under-the-hood change that switches currentLanguage
from a value managed in the tdash object to the native lang attribute of
HTML.

This means the updateLang function is no longer used externally and the
currentLang property no longer exists.

The old king is dead, long live the new king. Long live
`document.documentElement.lang`.

This is both your getter and setter from henceforth.